config.yml: No passwords in nested YAML structures
Work around a Dancer bug which overwrites passwords with "Hidden (looks
potentially sensitive)" when they are stored anywhere else than the
configuration file's top level.

This bug breaks the application after a runtime error has been encountered and
required a restart for it to be fixed.

Move the SNCF API credentials to the top level of the configuration file.