Add algorithm to guess lines for certain trains
When a given train cannot be found in the database (for example when it wasn't
originally scheduled, or due to circumstances such as strikes), it is usually
impossible to know which line it serves.

However, it can be usually inferred by its mission code and the station at
which it terminates.

This commit introduces a little algorithm that is used in such circumstances to
obtain the line anyway.