Normalize train numbers
Sometimes the Transilien API returns train numbers that aren't really a number,
e.g. "145312-145313".

The trains that have numbers like these usually traverse Paris, where they
change numbers because odd/even numbers indicate a certain direction to/from
Paris.  Problem is, there are a lot of trains that do this.

In this case, we only take the even number, because that's easier.