DB: fix missing trains in station_next_trains
Rewrite the station_next_trains stored procedure.

That stored procedure was incorrect because it was missing services that
normally wouldn’t run on a given day, but were included as an exception with
calendar_dates. Also, there was some very ugly copy/pasting going on.

As an extra benefit, the rewritten function runs significantly faster than the
previous version, with an eight- to tenfold reduction in execution time (from
1 second to 0.1-0.125 s).