Cleanup: use Dancer::Plugin::Database
Let Dancer connect to the database and pass the handle around in the rest of
the code.

Not only is it a pity not to use Dancer’s ability to handle database
connections, but it should also help reduce the number of concurrent
connections to said database.

I haven’t tested it, but this change may also have the added benefit of
reducing the number of open file descriptors and fix a file descriptor leak
which, in production, required me to restart my websites regularly, because
they broke due to an exhaustion of file descriptors.