UI: force line break between code and number
We used to rely on the natural overflow/line break in order to force a break
between the train’s name and number. However, this doesn’t always work;
especially when the train’s code is very short. A good example is the train
named “ICLI” and numbered “ICEL13”.

This line break must however not be displayed on phones.