Migrate and rewrite installer; switch GTFS formats
This is the first commit in a series that will migrate the project from both
MySQL and Redis to PostgreSQL, also getting rid of Redis in the process.

This commit rewrites the installer script so as to behave like a true
installer script: it creates the database (but you still need to create the
roles yourself to give them passwords), all tables and fills them
appropriately. Overall, the structure of the installer script is improved and
some ugly warts removed.

The database objects are created using an new intermediate role that has more
rights than the normal day-to-day database role, without being the
superuser. This enables a later update script to use only the privileges of
that updater role, therefore safely allowing updates using a daily cron job
and a .pgpass file.

This commit also switches to SNCF’s new GTFS format. Experimenting directly
with the new GTFS exports revealed that the preprocessing scripts were no
longer necessary (actually, it is debatable whether these scripts were needed
at all). Instead, we define the GTFS tables directly in our SQL files.

What’s now left to do is to distinguish between fresh installs and updates.