Migrate fetching full train schedule
Migrate the second case, when real-time data is unavailable. In this case, we
fall back to the fixed schedule in GTFS form.

This will come with a significant performance boost because the underlying
query was completely rewritten in order to benefit fully from PostgreSQL’s
capabilities. The equivalent MySQL query took 1 s and this query now takes
about 20 ms, which is a 50-fold speedup.