Migrate caching from Redis to PostgreSQL
PostgreSQL’s ability to store JSON data as JSONB type is interesting because
it means that PostgreSQL can not only replace MySQL, but also Redis.

We therefore need to create a table and some stored procedures that can be
easily called from Perl.

For the time being, the caching is left pretty much as is, but a future commit
may implement a better strategy.