Add update script
Write a separate update script. The script should be faster and update
only the necessary tables without requiring superuser privileges.

The installer script takes care to create all database objects as a
lesser-privileged user, which means that the updater script can be run with a
dedicated updater role whose password could be stored in a .pgpass file,
making it suitable for cron jobs.