Properly handle trains with changing numbers
Trains on the D line change numbers between Paris Gare de Lyon and
Châtelet-Les Halles; odd numbers become even and vice-versa.

In order to correctly display those train numbers, we need to know where a
train is and whether it is traveling from or towards Paris. This is
accomplished by extra columns in the station_lines table, assigning an index
number to a station on a given line and telling whether increasing index
numbers means traveling from or towards Paris.

Fortunately, no other line requires this, although it is quite possible that
the RER E may need this too when its extension to the west is complete.