[#20] Use Transilien API for some RER B stations
Some stations served by the RER A and B lines can, in fact, be queried using
the Transilien API, yet the site would only display the theoretical timetable
for these stations. This is corroborated by the API's specification document,
which lists the stations this API can be queried for.

This problem happened because an old version of my code assumed that any
station served by these lines could not get real-time train times at all.

Commit b7770eda133731576158d95fe926659aab80b0ca partially fixed the issue by
changing that logic: any station served by at least one SNCF line (RER C, D, E
or any of the Transilien lines) is assured to be eligible and all other
stations are assumed not to be. That meant that all SNCF-operated stations on
the RER A were fixed because they are also served by Transilien J or L.

However, all SNCF-operated stations on the RER B without connections to SNCF
lines were still left out.

Fixing this long-standing issue required listing explicitly all those
SNCF-operated stations in the database, by introducing an extra column which
gives a hint about which data source to use for each station.

These hints may need to be moved to a separate table later if more complex
decisions need to be made on the order in which data sources should be queried
for a given station. But for now, this is sufficient.

The affected stations were all on the RER B line:

 * Aéroport Charles de Gaulle 1
 * Aéroport Charles de Gaulle 2 TGV
 * Drancy
 * La Courneuve - Aubervilliers
 * La Plaine - Stade de France
 * Le Blanc-Mesnil
 * Le Bourget
 * Parc des Expositions
 * Sevran Beaudottes
 * Sevran - Livry
 * Vert Galant
 * Villeparisis - Mitry-le-Neuf
 * Villepinte

In order to fix this bug in production:

 * run the migration script as the database superuser:
   install/migrate/20220128_station_names_add_transilien_ok_column.sql

 * run update.sh in order to reload the station table with the correct data.

This commit fixes issue #20.