Fix subtle autocompletion failures
The station name autocompletion feature could fail in certain ways and in some
database configurations.

For example, searching for "Massy" failed to return any result, even though
there are two stations whose names start with this string. This is due to the
fact these stations have names starting with "Massy", followed by a
non-breaking space, an en-dash and another non-breaking space. The bug was not
reproducible on my development machine in the first place. It turns out the
culprit was a wrong LC_CTYPE setting for the database, so the install script
now explicitly sets LC_COLLATE and LC_CTYPE at database creation. Production
databases affected by this problem need to be dropped and recreated.

Another problem example: searching for "Er" should return at least two
stations (Ermont-Halte and Ermont-Eaubonne). Because of a to_tsquery function
call failing to specify the fr_unaccent configuration, the string "Er" was
treated as a stop word and ignored. This commit also fixes this issue.