Improve suggested stations when autocompleting
Ever since migrating from MySQL to PostgreSQL, I wanted to try using
PostgreSQL's more advanced full-text search (FTS) capabilities for station
name completion.

However, the ranking of the suggestions was still unsatisfactory. That is
because PostgreSQL’s built-in functions for ranking work best in a slightly
different use case than ours, i.e. searching text inside larger documents,
whereas here, they are used for suggesting names. After many hours of
experimentation, I came to the conclusion that a good ranking algorithm would
rank the results in the following order: complete matches first, then prefix
matches, then matches starting anywhere. The new rewritten query does exactly
that.

Searches are performed on the station's code and name, which means that a
string of "cl" will yield Creil as its first result because its code is "CL";
then all stations whose name or code start with "cl" follow; then all stations
which contain a word starting with "cl" anywhere but at the beginning are
last.

Station names also need to be cleaned up a bit, because hyphens in names
should be treated as word breaks and the simple text search configuration does
not yield satisfactory results in this case. Writing a preprocessing function
in SQL and running queries and stations names through it before converting
those strings to tsvectors or tsqueries is crude but works as it should.

So with all this work, station searching should be much closer to what I
believe is correct. One small possible improvement would be to use historical
ridership data, or some other kind of “popularity contest”-type of ranking, as
an extra input to the ranking algorithm in order to break ties.