Ensure consistent names for RER A and B trains
Trains on RER lines A and B have a number which actually is its proper name
followed by two digits. In other words, the four first characters of a train
“number” on these lines should be equal to its train name as recorded in the
database.

However, pulling train data from the database for such lines sometimes yielded
trains whose names did not match with the first four characters of the number.
For example, a train named “UZAR” but whose number is “UZEL55” really should
be named “UZEL” instead. This can be confirmed by looking at the real-time
data as supplied by SNCF, or the RATP’s website: both sources name the same
train “UZEL”.

This commit introduces a function which takes a train name and a train number.
If the train number is in the RATP style, then it returns its first four
characters. If not, it returns the train name untouched. It is then used in
the query for the next trains according to the database.