Harden JavaScript code against DOM-XSS
The (client-side) JavaScript code in the webpage uses Element.innerHTML to
refresh information dynamically. This is bad.

Using innerHTML has obvious security risks. Suppose a malicious actor directs
a victim to a malicious version of the site, where the /json API endpoint
would inject arbitrary HTML in the results. Because the data is fed to
Element.innerHTML without any kind of filtering, the victim would be
vulnerable to XSS attacks.

All calls to Element.innerHTML are now replaced by a combination of creating
DOM elements with document.createElement() and the use of Node.textContent.