Improve/fix error cases when checking station codes
Some stations, whose codes are not necessarily known, get assigned virtual
codes starting with NC or ND followed by a digit. Selecting some of these
stations led to a “Station not found” error, even though the station does
exist: this was due to the RER::Web::check_code() function being overzealous.

Furthermore, the station preference cookie was set in JavaScript before
checking that the station actually exists. The call to the set_cookie()
function is moved to the function processing an incoming set of data, to
ensure that this only happens in a successful code path.

Finally, when reading the cookie, the server did not check if it pointed to a
valid station before issuing the 302 redirection. This could get a user stuck
in a situation where visiting the root URL of the site would forcefully
redirect it to a 404 error. The only way the user could get himself out of
this is by deleting the cookie.

Now, the server checks the station actually exists and if not, disregards the
cookie.

This fix does, however, introduce a very slight performance penalty, because
the Perl code fetches the same database row twice; however the stations table
is indexed by code and one can reasonably expect the second fetch to be a
cache hit.