Add priorities to messages at bottom of screen
The messages framework receives a new feature: each message can now have its
own priority.

Eachc message priority is mapped to different color schemes that I’ve seen in
use on Infogare screens: white on a dark background for low priority messages,
white on a purple background, very dark blue on a yellow background with a
blinking dot on the left, and for the worst disruptions, white on a red
background.

Surprisingly, the code that would allow us to use different priorities was
already ready server-side, but was never used until now. So most of the
changes are made client-side.

Eventually, this system could be used to display internal error messages more
elegantly than what we are doing now. This has not been changed yet.

This commit also introduces another minor cosmetic change with respect to the
area at the bottom of the page: even when there are no messages, the adequate
space is now reserved in order to avoid the footer jumping around when
messages come and go.