Bugfix: links were inoperative after one click
Due to the body of the autocomp_reset() function being wrong, the links on the
footer of the page only worked if they were the first thing to be clicked.

The correct behavior for a click event handler at the document level is the
following: if the autocomplete box is open, close it, reset its contents to
the current station and stop performing any further actions. If not, let the
event be handled normally. The rewritten function now behaves as described.