Use correct UIC8 codes for Transilien API requests
Commit d400e6b introduced a column named “transilien_api_ok” to signal which
stations could be requested through the Transilien API.

Stations can have one or more UIC codes, which are used as search keys for
this API. Most of the time, all UIC codes for the same station work equally
well for requests.

However, there are a few exceptions: some stations have more than one UIC code
in the database, but only one of these is accepted by the API. This is namely
the case of Massy Verrières, Massy Palaiseau and La Défense.

For these exceptions, it was clear that the current approach would not work.
Instead, I imported the list of UIC8 codes accepted by the Transilien API from
the manual (there is, however, a mistake in the UIC8 code for Trappes which I
corrected in this dataset and reported to SNCF), dropped the transilien_api_ok
column in station_names and rewrote the find_station_by_key() stored procedure
to return the correct UIC8 code to use for the Transilien API, if it exists.

This also means that there is no need to compute the UIC8 code by ourselves
anymore, because we are pulling the 8-digit codes directly from the database.
The corresponding Perl code is therefore removed.