PRIM: Determine if train is stopping
PRIM returns all trains passing at a station, including those that do not
stop. We need to be able to filter them out in the results.

The implementation is a bit of guesswork, based on experiments. RATP does not
seem to return trains that do not stop at a given station, but SNCF does, so
the code only works for SNCF.