Fix bashism in install.sh
Use of the “read” built-in command without any arguments only works in GNU
bash (to which /bin/sh is often symlinked on GNU/Linux systems). But “read” in
FreeBSD’s sh(1) expects at least one argument.

The only place where “read” is called in the script is to provide a
rudimentary means of confirming some settings before committing them to the
database. We need some variable name that we never read from. GNU bash assigns
to $REPLY if no variable name is given, so we know $REPLY is safe to use as
dummy variable to read into.