Database: add various indices where needed
In production, the website is now facing over 40,000 monthly users and the
machine the website is running on is grinding to a halt.

It seemed that PostgreSQL was being the bottleneck. Based on output from
EXPLAIN ANALYZE, it turned out that some queries needed indices in order to
avoid expensive sorting or searches.

On a test setup, a two-minute load test with siege and a scenario constructed
from an especially recent busy day in production leads to a 3.5-fold increase
in number of requests served, from 1,200 to 4,200 requests served.