Deprecate use of UIC codes in codebase
UIC codes were only used as search keys for the now-retired Transilien
real-time API. The migration to PRIM, which uses completely different codes,
means that we no longer need to maintain the ability to search for stations by
UIC code.

Tables and columns that were relevant back when the Transilien API was in
used, such as “valid_transilien_api_uics” or the “uic” column in
“station_codes”, are removed.

The “RER::Gare” class also loses the “uic” and “transilien_api_search_key”
attributes, which were both unused or filled with dummy values which the rest
of the code didn’t care about.

At the SQL side, many server-side functions are removed and
find_station_by_key() is rewritten, eliminating a JOIN. This means that there
might be a slight performance improvement.