Avoid UI jumping around
Due to a small programming oversight, whenever the user closed the “extra
options” drawer, it could reopen by itself because of the data being
refreshed. We make sure this only happens if the user changes stations. It is
useful to have the drawer open by default in this case and if it’s useful to
filter by lines, but if the user wants to keep it closed, it should stay closed.