Memoize station lookup
The only relatively expensive function in the pipeline of looking up the data
to serve to a user is to search the database for the station matching a given
StopPoint or StopArea.

While the database side of things is already well-optimized with indices, we
can squeeze even more performance by memoizing these searches on the server
side. This way, we avoid calling out to the database repeatedly for the same
station all the time.

The cache is invalidated every time a query is processed, so that updates to
the station database are effective at the next request.