GTFS: do not load unnecessary tables in database
Tables named agency, stop_extensions and transfers are not used by this
application. Besides, the agency table has always been a source of headaches
because the schema in the input CSV files changed constantly without any
notice, causing update scripts to fail. Not bothering with those tables will
help me out a lot.

We also do not use any columns at all in the “stops” table, except for the
stop_id in joins. We can get rid of this table too.